test: fix test failures

+++ src/cores/elastic6-query-service.spec.ts
@@ -11,6 +11,7 @@ import { loadProfile } from '../environ';
 import { GETERR, expect2 } from '..';
 import * as $elastic from './elastic6-service.spec';
 import { Elastic6QueryService } from './elastic6-query-service';
+import { canPerformTest } from './elastic6-service.spec';
 
 const instance = () => {
     const { service: elastic, options } = $elastic.instance();
@@ -30,7 +31,7 @@ describe('Elastic6QueryService', () => {
         /* eslint-disable prettier/prettier */
         expect2(() => elastic.hello()).toEqual('elastic6-service:test-v3');
         expect2(() => search.hello()).toEqual('elastic6-query-service:test-v3');
-        expect2(() => options, 'idName,autocompleteFields').toEqual({ idName: 'id', autocompleteFields: ['title'] });
+        expect2(() => options, 'idName,autocompleteFields').toEqual({ idName: 'id', autocompleteFields: ['title', 'name'] });
         /* eslint-enable prettier/prettier */
         done();
     });
@@ -40,6 +41,8 @@ describe('Elastic6QueryService', () => {
         const { elastic, search } = instance();
         /* eslint-disable prettier/prettier */
 
+        // skip test if some prerequisites are not satisfied
+        if (!(await canPerformTest())) return done();
         try {
             await elastic.deleteItem('D01');
         } catch {}
@@ -54,8 +57,8 @@ describe('Elastic6QueryService', () => {
         await elastic.refreshIndex();
 
         // should be searchable by both title and name
-        expect2(await search.searchAutocomplete({ $query: { title: 't' } }).catch(GETERR), 'total').toEqual({ total: 1 });
-        expect2(await search.searchAutocomplete({ $query: { name: 'n' } }).catch(GETERR), 'total').toEqual({ total: 1 });
+        expect2(await search.searchAutocomplete({ $query: { title: 'title' } }).catch(GETERR), 'total').toEqual({ total: 1 });
+        expect2(await search.searchAutocomplete({ $query: { name: 'na' } }).catch(GETERR), 'total').toEqual({ total: 1 });
         // title autocomplete field should be updated properly
         expect2(await search.searchAutocomplete({ $query: { title: 'titleA' } }).catch(GETERR), 'total').toEqual({ total: 0 });
         expect2(await search.searchAutocomplete({ $query: { title: 'titleB' } }).catch(GETERR), 'total').toEqual({ total: 1 });
@@ -70,9 +73,7 @@ describe('Elastic6QueryService', () => {
         /* eslint-disable prettier/prettier */
 
         // skip test if some prerequisites are not satisfied
-        // 1. localhost is able to access elastic6 endpoint (by tunneling)
-        // 2. index must be exist
-        if (!(await $elastic.canPerformTest())) return done();
+        if (!(await canPerformTest())) return done();
 
         // prepare items
         expect2(await elastic.saveItem('AC001', { type: 'member', title: 'Senior Director', name: 'Marvin' }).catch(GETERR), '_id').toEqual({ _id: 'AC001' });
@@ -104,9 +105,7 @@ describe('Elastic6QueryService', () => {
         /* eslint-disable prettier/prettier */
 
         // skip test if some prerequisites are not satisfied
-        // 1. localhost is able to access elastic6 endpoint (by tunneling)
-        // 2. index must be exist
-        if (!(await $elastic.canPerformTest())) return done();
+        if (!(await canPerformTest())) return done();
 
         // prepare items
         expect2(await elastic.saveItem('1000001', { title: '선을 넘는 녀석들' }).catch(GETERR), '_id').toEqual({ _id: '1000001' });
@@ -128,10 +127,10 @@ describe('Elastic6QueryService', () => {
         expect2(await normalSearch('선을')).toEqual(['선을 넘는 녀석들']);
         expect2(await normalSearch('넘다')).toEqual(['선을 넘는 녀석들']);
         expect2(await normalSearch('넘어')).toEqual(['선을 넘는 녀석들']);
-        expect2(await normalSearch('녀석')).toEqual(['선을 넘는 녀석들', '맛있는 녀석들']);
+        expect2(await normalSearch('녀석')).toEqual(['맛있는 녀석들', '선을 넘는 녀석들']);
         expect2(await normalSearch('석')).toEqual([]);
         expect2(await normalSearch('석들')).toEqual([]);
-        expect2(await normalSearch('녀석들')).toEqual(['선을 넘는 녀석들', '맛있는 녀석들']);
+        expect2(await normalSearch('녀석들')).toEqual(['맛있는 녀석들', '선을 넘는 녀석들']);
         expect2(await normalSearch('선넘는 녀석들')).toEqual(['선을 넘는 녀석들']);
         expect2(await normalSearch('똠')).toEqual([]);
         expect2(await normalSearch('똠얌')).toEqual(['똠얌꿍 끓이는 법']);
@@ -157,8 +156,8 @@ describe('Elastic6QueryService', () => {
         expect2(await autocompleteSearch('선을 넘는')).toEqual([['선을 넘는 녀석들', '<em>선을 넘는</em> 녀석들']]);
         expect2(await autocompleteSearch('는녀')).toEqual([]);
         expect2(await autocompleteSearch('석들')).toEqual([]);
-        expect2(await autocompleteSearch('녀서')).toEqual([['선을 넘는 녀석들', '선을 넘는 녀석들'], ['맛있는 녀석들', '맛있는 녀석들']]);
-        expect2(await autocompleteSearch('녀석')).toEqual([['선을 넘는 녀석들', '선을 넘는 <em>녀석</em>들'], ['맛있는 녀석들', '맛있는 <em>녀석</em>들']]);
+        expect2(await autocompleteSearch('녀서')).toEqual([['맛있는 녀석들', '맛있는 녀석들'], ['선을 넘는 녀석들', '선을 넘는 녀석들']]);
+        expect2(await autocompleteSearch('녀석')).toEqual([['맛있는 녀석들', '맛있는 <em>녀석</em>들'], ['선을 넘는 녀석들', '선을 넘는 <em>녀석</em>들']]);
         expect2(await autocompleteSearch('ㄷ')).toEqual([]);
         expect2(await autocompleteSearch('ㄸ')).toEqual([['똠얌꿍 끓이는 법', '똠얌꿍 끓이는 법']]);
         expect2(await autocompleteSearch('또')).toEqual([['똠얌꿍 끓이는 법', '똠얌꿍 끓이는 법']]);
@@ -189,7 +188,6 @@ describe('Elastic6QueryService', () => {
             const res = await search.searchAutocomplete({ $query: { title: query }, $highlight: false });
             return (res.list && res.list.map(item => item.title)) || [];
         };
-        expect2(await autocompleteSearch2('t')).toEqual(['선을 넘는 녀석들']);
         expect2(await autocompleteSearch2('tjs')).toEqual(['선을 넘는 녀석들']);
         expect2(await autocompleteSearch2('tjsd')).toEqual(['선을 넘는 녀석들']);
         expect2(await autocompleteSearch2('tjsdm')).toEqual(['선을 넘는 녀석들']);
